feat(!): new alias command for all types of packages
BREAKING CHANGE: New alias command that is to be used for creating package, npm and import map aliases. Old alias commands have been removed.
